requirements.txt
numpy==2.0.1
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
Pillow==10.1.0
Pillow==10.4.0
torch==2.1.2
tqdm==4.66.4
transformers==4.36.2
